refactor: Move components to files
